requirements_llama.txt
numpy==1.22.4
sentencepiece==0.1.97
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1
git+https://github.com/daniel-furman/transformers_llama@llama_push#egg=transformers
datasets==2.10.1
accelerate==0.16.0
bitsandbytes==0.37.0
tqdm==4.65.0
black==23.1.0
flake8==6.0.0
deep-translator==1.10.1
python-dotenv